--- src/main/java/com/todolist/dto/ErrorResponse.java
@@ -1,11 +1,0 @@
-package com.todolist.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ErrorResponse {
-    private String code;
-    private String message;
-}
