+++ src/main/java/com/todolist/exception/ErrorResponse.java
@@ -1,0 +1,20 @@
+package com.todolist.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.Map;
+
+@Getter
+@AllArgsConstructor
+public class ErrorResponse {
+    private String code;
+    private String message;
+    private Map<String, String> details;
+
+    public ErrorResponse(String code, String message) {
+        this.code = code;
+        this.message = message;
+
+    }
+}
